scripts/ingest/requirements.txt
supabase==2.27.2
requests==2.32.3
beautifulsoup4==4.14.3
lxml==5.3.0
python-dotenv==1.0.1
